# services/api-go/internal/handler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KineGuide AI Main API",
    "version": "0.2.0",
    "description": "Public Go API for the privacy-preserving KineGuide educational prototype. Raw camera frames, images, and video are explicitly excluded from every request."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Direct local development"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "registerAccount",
        "tags": [
          "Authentication"
        ],
        "summary": "Create an account and refresh session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; rotating refresh token is set as an HttpOnly cookie.",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "Authentication"
        ],
        "summary": "Sign in with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in; rotating refresh token is set as an HttpOnly cookie.",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refreshSession",
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate a refresh token and issue a new access token",
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Access token refreshed and refresh cookie rotated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke the refresh session",
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Refresh session revoked."
          }
        }
      }
    },
...
    "/v1/auth/oauth/{provider}/start": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OAuthProvider"
        }
      ],
      "get": {
        "operationId": "startOAuthLogin",
        "tags": [
          "Authentication"
        ],
        "summary": "Start authorization-code social sign-in",
        "description": "Sets a short-lived HttpOnly state cookie and redirects to the provider. No provider token is returned to the browser.",
        "responses": {
          "302": {
            "description": "Redirect to the configured provider authorization page.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/auth/oauth/{provider}/callback": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OAuthProvider"
        },
        {
          "name": "code",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "One-time authorization code from the provider."
        },
        {
          "name": "state",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Anti-forgery state returned by the provider."
        },
        {
          "name": "error",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Provider cancellation or failure code."
        }
      ],
      "get": {
        "operationId": "completeOAuthLogin",
        "tags": [
          "Authentication"
        ],
        "summary": "Validate the provider callback and start or link a KineGuide session",
        "responses": {
          "303": {
            "description": "Redirect to the fixed web callback with a safe success or error code. A successful login sets the KineGuide refresh cookie.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "Authentication"
        ],
        "summary": "Get the authenticated account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUser",
        "tags": [
          "Authentication"
        ],
        "summary": "Delete the account and all consent, assessment, token, identity, and session records",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Account and owned records deleted."
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/me/auth-identities": {
      "get": {
        "operationId": "listAuthIdentities",
        "tags": [
          "Authentication"
        ],
        "summary": "List social identities connected to the current account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Connected provider names and connection timestamps; provider subjects and tokens are excluded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "identities"
                  ],
                  "properties": {
                    "identities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuthIdentity"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/me/auth-identities/{provider}/start": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OAuthProvider"
        }
      ],
      "post": {
        "operationId": "startAuthIdentityLink",
        "tags": [
          "Authentication"
        ],
        "summary": "Start an authenticated social-account linking flow",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider authorization URL bound to the authenticated KineGuide user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/me/auth-identities/{provider}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OAuthProvider"
        }
      ],
      "delete": {
        "operationId": "disconnectAuthIdentity",
        "tags": [
          "Authentication"
        ],
        "summary": "Disconnect a social identity when another login method remains",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Social identity disconnected."
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/v1/consents": {
      "post": {
        "operationId": "recordConsent",
        "tags": [
          "Consent"
        ],
        "summary": "Record versioned consent choices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consent recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kg_refresh"
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "Safe error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          },
          "display_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "writeOnly": true
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1,
            "description": "Access-token lifetime in seconds."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
    }
  }
}